Add new FO page 'create account'
